+++ src/__playground__/index.ts
@@ -1,15 +1,6 @@
-import { Messenger } from '../index.js';
+import { ServiceCluster } from '../index.js';
 import { api } from './worker.js';
 
-const m1 = new Messenger('foo');
-const m2 = new Messenger(m1.transfer());
+const cluster = new ServiceCluster(api);
 
-m2.onMessage<string>((data) => {
-    console.log(data);
-});
-
-m1.sendMessage('foo');
-
-await new Promise((resolve) => setTimeout(resolve, 5e3));
-
-// await api({ name: 'seek' });
+const x = await cluster.launch(1, { reffed: true });